Analysis of Covid Stats in India
India fights Corona Covid-19

Coronavirus disease (COVID-19) is an infectious disease caused by the SARS-CoV-2 virus. Government of India is taking all necessary steps to ensure that we are prepared well to face the challenge and threat posed by the growing pandemic of COVID-19 the Corona Virus. Purpose of this project to  represent the state wise data of num of cases of corona, deaths, cured , vaccinated & how Corona has impacted the life .And the majors taken to get people vaccinated